+++ core/_/EndReason.cs
@@ -1,5 +1,6 @@
 ﻿namespace YAGrep {
     public enum EndReason {
+        StillRunning,
         Canceled,
         Interrupted,
         EmptyInput,