+++ test/LineTest.cs
@@ -5,28 +5,28 @@ namespace Grep.Test {
     public class LineTest {
         [Test]
         public void can_setup_lines() {
-            var singleLine = new Line(_singleLine, startIndex: 0, length: 26);
+            var singleLine = new Line(_singleLine).Update(startIndex: 0, length: 26);
 
             Assert.That(singleLine.Length, Is.EqualTo(26));
             Assert.That(singleLine.AsString(), Is.EqualTo("Lorem ipsum dolor sit amet"));
 
-            var multipleLines = new Line(_multipleLines, startIndex: 0, length: 26);
+            var multipleLines = new Line(_multipleLines).Update(startIndex: 0, length: 26);
 
             Assert.That(multipleLines.Length, Is.EqualTo(26));
             Assert.That(multipleLines.AsString(), Is.EqualTo("Lorem ipsum dolor sit amet"));
 
-            multipleLines = new Line(_multipleLines, startIndex: 27, length: 27);
+            multipleLines = new Line(_multipleLines).Update(startIndex: 27, length: 27);
             Assert.That(multipleLines.Length, Is.EqualTo(27));
             Assert.That(multipleLines.AsString(), Is.EqualTo("Consectetur adipiscing elit"));
 
-            multipleLines = new Line(_multipleLines, startIndex: 55, length: 32);
+            multipleLines = new Line(_multipleLines).Update(startIndex: 55, length: 32);
             Assert.That(multipleLines.Length, Is.EqualTo(32));
             Assert.That(multipleLines.AsString(), Is.EqualTo("Sed do eiusmod tempor incididunt"));
         }
 
         [Test]
         public void can_create_line_snapshot() {
-            var multipleLines = new Line(_multipleLines, startIndex: 27, length: 27);
+            var multipleLines = new Line(_multipleLines).Update(startIndex: 27, length: 27);
             var snapshot = multipleLines.Snapshot();
             Assert.That(multipleLines, Is.Not.EqualTo(snapshot));
             Assert.That(snapshot.AsString(), Is.EqualTo("Consectetur adipiscing elit"));
@@ -34,7 +34,7 @@ namespace Grep.Test {
 
         [Test]
         public void lenght_greater_than_buffer_is_invalid() {
-            var invalidLine = new Line(_singleLine, startIndex: 0, length: 27);
+            var invalidLine = new Line(_singleLine).Update(startIndex: 0, length: 27);
 
             Assert.That(invalidLine.Valid(), Is.False);
             Assert.That(invalidLine.Length, Is.EqualTo(-1));
@@ -43,7 +43,7 @@ namespace Grep.Test {
 
         [Test]
         public void can_search_chars() {
-            var line = new Line(_multipleLines, startIndex: 27, length: 27);
+            var line = new Line(_multipleLines).Update(startIndex: 27, length: 27);
             Assume.That(line.AsString(), Is.EqualTo("Consectetur adipiscing elit"));
 
             Assert.That(line.IndexOf('e', 0, 4), Is.EqualTo(-1));
@@ -54,7 +54,7 @@ namespace Grep.Test {
 
         [Test]
         public void can_reference_chars() {
-            var line = new Line(_multipleLines, startIndex: 27, length: 27);
+            var line = new Line(_multipleLines).Update(startIndex: 27, length: 27);
             Assume.That(line.AsString(), Is.EqualTo("Consectetur adipiscing elit"));
 
             Assert.That(line[2], Is.EqualTo('n'));
@@ -63,12 +63,11 @@ namespace Grep.Test {
 
         [Test]
         public void should_enforce_line_boundaries_on_char_search() {
-            var line = new Line(_multipleLines, startIndex: 27, length: 27);
+            var line = new Line(_multipleLines).Update(startIndex: 27, length: 27);
             Assume.That(line.AsString(), Is.EqualTo("Consectetur adipiscing elit"));
 
             Assert.That(line.IndexOf('e', 24, 50), Is.EqualTo(-1));
         }
-
 
         #region possibilities
         private readonly char[] _singleLine = "Lorem ipsum dolor sit amet".ToCharArray();