+++ CHANGELOG.md
@@ -9,9 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - case sensitivity
 - regex support
-- count matches
-- include line number on result
 - multiple files and directories
+
+## [0.0.2] - 2021-05-18
+
+### Added
+
+- Trim support
+- Include line number on result
+- Count total matches and runtime on console app
 
 ## [0.0.1] - 2021-05-05